Move converters out of utils
