+++ parcels/utils/__init__.py
@@ -1,2 +1,1 @@
-from .converters import *
 from .interpolation_utils import *
